requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
psycopg2-binary==2.9.11
SQLAlchemy==2.0.49
pandas==2.2.1
python-dotenv==1.2.1
google-generativeai==0.4.1
requests==2.31.0
